Consolidate all settings into workspace settings
