--- packages/stryker-api/.vscode/settings.json
@@ -1,14 +1,0 @@
-{
-  "files.exclude": {
-    ".git": true,
-    "**/*.js": {
-      "when": "$(basename).ts"
-    },
-    "**/*.d.ts": {
-      "when": "$(basename).ts"
-    },
-    "**/*.map": {
-      "when": "$(basename)"
-    }
-  }
-}
