--- packages/stryker-babel-transpiler/.vscode/settings.json
@@ -1,13 +1,0 @@
-{
-  "files.exclude": {
-      ".git": true,
-      ".tscache": true,
-      "**/*.js": {
-        "when": "$(basename).ts"
-      },
-      "**/*.d.ts": true,
-      "**/*.map": {
-        "when": "$(basename)"
-      }
-  }
-}
