--- packages/stryker-html-reporter/.vscode/settings.json
@@ -1,17 +1,0 @@
-{
-  "files": {
-    "exclude": {
-      ".git": "",
-      "**/*.js": {
-        "when": "$(basename).ts"
-      },
-      "**/**.js": {
-        "when": "$(basename).tsx"
-      },
-      "**/*.d.ts": true,
-      "**/*.map": {
-        "when": "$(basename)"
-      }
-    }
-  }
-}
