--- packages/stryker-jasmine-runner/.vscode/settings.json
@@ -1,12 +1,0 @@
-{
-  "files.exclude": {
-    ".git": true,
-    "**/*.js": {
-      "when": "$(basename).ts"
-    },
-    "**/*.d.ts": true,
-    "**/*.map": {
-      "when": "$(basename)"
-    }
-  }
-}
