--- packages/stryker-jasmine-runner/typings/JasmineTypes.js
@@ -1,2 +1,0 @@
-"use strict";
-//# sourceMappingURL=JasmineTypes.js.map
