--- packages/stryker-jasmine/.vscode/settings.json
@@ -1,14 +1,0 @@
-{
-  "files": {
-    "exclude": {
-      ".git": "",
-      "**/*.js": {
-        "when": "$(basename).ts"
-      },
-      "**/*.d.ts": true,
-      "**/*.map": {
-        "when": "$(basename)"
-      }
-    }
-  }
-}
