+++ packages/stryker-jest-runner/tsconfig.src.json
@@ -5,7 +5,7 @@
   },
   "include": [
     "src",
-    "types"
+    "typings"
   ],
   "references": [
     {
