+++ packages/stryker-jest-runner/tsconfig.test.json
@@ -8,7 +8,7 @@
   },
   "include": [
     "test",
-    "types"
+    "typings"
   ],
   "references": [
     {
