+++ packages/stryker-jest-runner/typings/jest/index.d.ts
