--- packages/stryker-karma-runner/typings/karmaTypes.js
@@ -1,3 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-//# sourceMappingURL=karmaTypes.js.map
