--- packages/stryker-karma-runner/typings/karmaTypes.ts
@@ -1,14 +1,0 @@
-import * as karma from 'karma';
-
-declare module 'karma' {
-  interface Config extends karma.ConfigOptions { }
-
-  interface ConfigOptions {
-    detached?: boolean;
-    configFile?: string;
-  }
-
-  interface Reporter {
-    adapters: any[];
-  }
-}
