--- packages/stryker-mocha-framework/.vscode/settings.json
@@ -1,14 +1,0 @@
-{
-  "files.exclude": {
-    "**/.git": true,
-    "**/.svn": true,
-    "**/.hg": true,
-    "**/CVS": true,
-    "**/.DS_Store": true,
-    "**/*.map": true,
-    "**/*.d.ts": true,
-    "**/*.js": {
-      "when": "$(basename).ts"
-    }
-  }
-}
