--- packages/stryker-mocha-runner/.vscode/settings.json
@@ -1,16 +1,0 @@
-{
-  "files": {
-    "exclude": {
-      ".git": "",
-      "**/*.js": {
-        "when": "$(basename).ts"
-      },
-      "**/*.d.ts": {
-        "when": "$(basename).ts"
-      },
-      "**/*.map": {
-        "when": "$(basename)"
-      }
-    }
-  }
-}
