--- packages/stryker-vue-mutator/.vscode/settings.json
@@ -1,14 +1,0 @@
-{
-  "typescript.tsdk": "../../node_modules/typescript/lib",
-  "files.exclude": {
-      ".git": true,
-      ".tscache": true,
-      "**/*.js": {
-        "when": "$(basename).ts"
-      },
-      "**/*.d.ts": true,
-      "**/*.map": {
-        "when": "$(basename)"
-      }
-  }
-}
