--- packages/stryker-wct-runner/.vscode/settings.json
@@ -1,15 +1,0 @@
-{
-  "files.exclude": {
-    ".git": true,
-    ".tscache": true,
-    "**/*.js": {
-      "when": "$(basename).ts"
-    },
-    "**/*.d.ts": false,
-    "test/**/*.d.ts": true,
-    "src/**/*.d.ts": true,
-    "**/*.map": {
-      "when": "$(basename)"
-    }
-  }
-}
