--- packages/stryker-webpack-transpiler/typings/enhanced-resolve.js
@@ -1,2 +1,0 @@
-"use strict";
-//# sourceMappingURL=enhanced-resolve.js.map
