--- packages/stryker/.vscode/settings.json
@@ -1,13 +1,0 @@
-{
-  "files.exclude": {
-    ".git": true,
-    ".tscache": true,
-    "src/*.js": {
-      "when": "$(basename).ts"
-    },
-    "{src,test}/*.d.ts": true,
-    "**/*.map": {
-      "when": "$(basename)"
-    }
-  }
-}
